Allow Direct3D hardware scaling to be toggled
Direct3D hardware scaling renders things at whatever resolution the
screen is set to, not the resolution of the game. There might be a way
to do Direct3D hardware scaling the right way, but there's already a
code path that exists to do scaling properly through software. This
provides access to Allegro-based software scaling in Direct3D for people
who prefer it (at the cost of speed).