Scripting: Do not parse *= as one entity in parameter declarations
This handles *= as a special case inside a parameter declaration. I
wasn't aware of this initially, but the inclusion of a *= symbol breaks
compatibility with one particular syntax for parameter declarations
available in older versions of AGS. The symbol is now broken into two
when it occurs directly after a variable type.